move server config to root
